Added deleteByIdUP() to BeerControllerIT

+++ src/main/java/guru/springframework/spring6restmvc/services/BeerServiceJPA.java
@@ -54,8 +54,13 @@ public class BeerServiceJPA implements BeerService {
     }
 
     @Override
-    public void deleteById(UUID uuid) {
-        beerRepository.deleteById(uuid);
+    public boolean deleteById(UUID uuid) {
+        if (beerRepository.existsById(uuid)) {
+            beerRepository.deleteById(uuid);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     @Override
